added the simple-input / output interfaces, cleaned up descriptions
